requirements.txt
flask
python-dotenv
openai
flask_sqlalchemy
bcrypt
psycopg2-binary